delete old generation method
