+++ packages/model/src/data/index.ts
@@ -6,7 +6,6 @@
  */
 
 export { default as generateCombinations, GenerateSetCombinationsOptions } from './generateCombinations';
-export { default as generateCombinations2 } from './generateCombinations2';
 export { default as asSets, asSet, PostprocessSetOptions } from './asSets';
 export {
   default as asCombinations,
@@ -19,6 +18,3 @@ export { default as extractSets } from './extractSets';
 
 export * from './setOverlap';
 export { default as setOverlap } from './setOverlap';
-
-export { PowerSetOptions } from './powerSet';
-export { default as powerSet } from './powerSet';
